refactored fault injection so that the product side has minimal surface
